// partners/partners-app/src/pages/list-products-page/list-products/list-products.types.tsx
export type ProductListingDeliveryTarget = 'HOST' | 'VENUE';

export interface ProductListingValues {
  is_duncit_delivery_partner: boolean;
  product_name: string;
  image_urls: string[];
  description: string;
  size_label: string;
  height_cm: number | string;
  weight_kg: number | string;
  color: string;
  inventory_count: number | string;
  unit_cost: number | string;
  commission_pct: number;
  delivery_target: ProductListingDeliveryTarget;
}